Fix clippy and move struct in meilisearch-type
